Move test suites into top-level directory
